requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
cachetools==4.0.0         # via google-auth
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
click==7.0                # via flask
dominate==2.4.0           # via flask-bootstrap
flask-bootstrap==3.3.7.1
flask-cors==3.0.8
flask-wtf==0.14.3
flask==1.1.1
google-api-core[grpc]==1.16.0  # via google-cloud-core, google-cloud-datastore
google-auth==1.11.0       # via google-api-core
google-cloud-core==1.3.0  # via google-cloud-datastore
google-cloud-datastore==1.10.0
googleapis-common-protos==1.51.0  # via google-api-core
grpcio==1.27.1            # via google-api-core
idna==2.8                 # via requests
itsdangerous==1.1.0       # via flask, flask-wtf
jinja2==2.11.1            # via flask
markupsafe==1.1.1         # via jinja2
protobuf==3.11.3          # via google-api-core, googleapis-common-protos
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
python-http-client==3.2.5  # via sendgrid
pytz==2019.3              # via google-api-core
requests==2.22.0          # via google-api-core
rsa==4.0                  # via google-auth
sendgrid==6.1.2
six==1.14.0               # via flask-cors, google-api-core, google-auth, grpcio, protobuf
urllib3==1.25.8           # via requests
visitor==0.1.3            # via flask-bootstrap
werkzeug==0.16.1          # via flask
wtforms==2.2.1            # via flask-wtf

# The following packages are considered to be unsafe in a requirements file:
# setuptools
